Add optional attribute to emit event while sliding the cursor
The consumer of this component can listen for the `range-changed` event
when the user has adjusted the ranges.

By default the event triggers only when the change has been completed,
that is when the sliding movement has finished.

In some cases the consumer may want to listen for changes also while
sliding the cursor, for example when using custom labels.

This is now possible by setting the `emit-event-on-update` attribute.